final order method and see order method and imrpovations

+++ EcommerceApplication/Service/OrderProcessorService.cs
@@ -12,7 +12,7 @@ using System.Xml.Serialization;
 
 namespace EcommerceApplication.Service
 {
-    public  class OrderProcessorService
+    public class OrderProcessorService
     {
         private readonly OrderProcessorRepositoryImpl OPRI;
         public OrderProcessorService() {
@@ -24,11 +24,15 @@ namespace EcommerceApplication.Service
         {
             try
             {
-                InvalidProductException.InvalidProductData(product);
+
+                if (product.Price < 0)
+                {
+                    throw new InvalidProductException("Invalid Price amount");
+                }
                 OPRI.CreateProduct(product);
                 Console.WriteLine("Record inserted successfully");
             }
-            catch (Exception ex)
+            catch (InvalidProductException ex)
             {
                 Console.WriteLine(ex.Message);
             }
@@ -39,11 +43,15 @@ namespace EcommerceApplication.Service
 
             try
             {
-                InvalidCustomerException.InvalidCustomerData(customer);
+                if (!customer.Email.Contains('@'))
+                {
+                    throw new InvalidCustomerException("Invalid Email");
+                }
                 OPRI.CreateCustomer(customer);
                 Console.WriteLine("Customer inserted successfully");
+                OPRI.GetRecentcustID();
             }
-            catch (Exception ex)
+            catch (InvalidCustomerException ex)
             {
                 Console.WriteLine(ex.Message);
             }
@@ -53,11 +61,14 @@ namespace EcommerceApplication.Service
         {
             try
             {
-                ProductNotFoundException.productnotfound(product);
+
+                if (!OPRI.ProductNotExist(product))
+                    throw new ProductNotFoundException("Product not found!!!");
+
                 OPRI.DeleteProduct(product);
                 Console.WriteLine("Product deleted");
             }
-            catch (Exception ex)
+            catch (ProductNotFoundException ex)
             {
                 Console.WriteLine(ex.Message);
             }
@@ -66,11 +77,12 @@ namespace EcommerceApplication.Service
         {
             try
             {
-                CustomerNotFoundException.customernotfound(customerid);
+                if (!OPRI.CustomerNotExist(customerid))
+                    throw new CustomerNotFoundException("Customer not found!!!");
                 OPRI.DeleteCustomer(customerid);
                 Console.WriteLine("Customer deleted");
             }
-            catch (Exception ex)
+            catch (CustomerNotFoundException ex)
             {
                 Console.WriteLine(ex.Message);
             }
@@ -82,13 +94,18 @@ namespace EcommerceApplication.Service
 
             try
             {
-                
-                ProductNotFoundException.productnotfound(product.ProductId);
-                ProductNotFoundException.NotEnoughStock(quantity,product.ProductId);
+
+                if (!OPRI.ProductNotExist(product.ProductId))
+                    throw new ProductNotFoundException("Product not found!!!");
+                if (quantity > OPRI.Availablestockquantity(product.ProductId))
+                {
+                    throw new ProductNotFoundException("Not Enough stock");
+                }
+
                 OPRI.AddToCart(customer, product, quantity);
                 Console.WriteLine("Product added to cart ");
             }
-            catch (Exception ex)
+            catch (ProductNotFoundException ex)
             {
                 Console.WriteLine(ex.Message);
             }
@@ -98,22 +115,28 @@ namespace EcommerceApplication.Service
         {
             try
             {
-                CustomerNotFoundException.customernotfound(customer.CustomerId);
-                ProductNotFoundException.productnotincart(product.ProductId,customer.CustomerId);
+                if (!OPRI.CustomerNotExist(customer.CustomerId))
+                    throw new CustomerNotFoundException("Customer not found!!!");
+                if (!OPRI.ProductNotExistinCart(product.ProductId, customer.CustomerId))
+                    throw new ProductNotFoundException("Product not found in cart for the customer!!!");
                 OPRI.RemoveFromCart(customer, product);
                 Console.WriteLine("Order deleted from cart ");
             }
-            catch (Exception ex)
+            catch (CustomerNotFoundException ex)
             {
+                Console.WriteLine(ex.Message);
+            }
+            catch (ProductNotFoundException ex) {
                 Console.WriteLine(ex.Message);
             }
 
         }
-       
+
         public void DisplayCartRecordservice(Customer customerid)
         {
-            try { 
-            CustomerNotFoundException.customernotfound(customerid.CustomerId);
+            try {
+                if (!OPRI.CustomerNotExist(customerid.CustomerId))
+                    throw new CustomerNotFoundException("Customer not found!!!");
                 List<Cart> cartList = OPRI.GetAllFromCart(customerid);
                 Console.WriteLine("cartid\tcustomerid\tproductid\tquantity");
                 foreach (Cart cart1 in cartList)
@@ -122,47 +145,44 @@ namespace EcommerceApplication.Service
                 }
 
             }
-            catch (Exception ex)
+            catch (CustomerNotFoundException ex)
             {
                 Console.WriteLine(ex.Message);
             }
         }
-        
-        public bool Placeorderservice(Customer customer, Order order, OrderItem orderitem, Product product)
+
+        public bool PlaceOrderservice(Customer customer, List<Dictionary<Product, int>> productsAndQuantities, string shippingAddress)
         {
             try
             {
-                CustomerNotFoundException.customernotfound(customer.CustomerId);
-                ProductNotFoundException.productnotfound(product.ProductId);
-                ProductNotFoundException.NotEnoughStock(orderitem.Quantity, product.ProductId);
-                OPRI.PlaceOrder(customer, order, orderitem, product);
-                Console.WriteLine("Order placed successfully");
+                if (!OPRI.CustomerNotExist(customer.CustomerId))
+                    throw new CustomerNotFoundException("Customer not found!!!");
+                OPRI.PlaceOrder(customer, productsAndQuantities, shippingAddress);
             }
-            catch (Exception ex)
+            catch (CustomerNotFoundException ex)
             {
-                Console.WriteLine(ex.Message);
+                Console.WriteLine(ex);
             }
             return true;
+
         }
-
-        public void DisplayOrderbyCustomer(Order order)
+        public void GetOrdersByCustomerservice(int customerId)
         {
-
             try
             {
-                //OrderNotFoundException.ordernotfound(order.OrderId);
-                List<OrderItem> orderList = OPRI.GetOrderByCustomer(order);
-                Console.WriteLine("Productid\tQuantity");
-                foreach (OrderItem list in orderList)
-                {
-                    Console.WriteLine(list.ProductId + "\t\t" + list.Quantity);
-                }
+                if (!OPRI.OrderNotExist(customerId))
+                    throw new OrderNotFoundException("Order not found fro the customer");
+                OPRI.GetOrdersByCustomer(customerId);
             }
-            catch(Exception ex)
+            catch (OrderNotFoundException ex)
             {
-                Console.WriteLine(ex.Message);
+                Console.WriteLine(ex);
             }
         }
+
+        
+
+
 
     }
